+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Support for variable declarations.
 - Support for struct types.
+- Support for packed arrays.
 
 ### Fixed
 - Fix blocking assignments, making them actually block.